add flask server
