+++ data/input.csv
@@ -1,3 +1,7 @@
-監視カメラ流用
-動線分析
-来店人数計測
+persol
+Panasonic
+Toyota
+recruit
+EPSON
+日本電気
+日揮
