+++ pyproject.toml
@@ -15,6 +15,7 @@ isort = "5.10.1"
 black = "21.10b0"
 flake8 = "4.0.1"
 mypy = "0.910"
+Flask = "^2.0.3"
 
 [tool.poetry.dev-dependencies]
